# services/auth-service/openapi.yaml
openapi: 3.0.3
info:
  title: Auth Service
  description: Handles user authentication, registration, and token management.
  version: 1.0.0
paths:
  /register:
    post:
      summary: Register a new user
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                username:
                  type: string
                  format: email
                  description: The user's email address.
                password:
                  type: string
                  format: password
                  description: The user's password.
              required:
                - username
                - password
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Invalid input
  /token:
    post:
      summary: Get JWT token for user
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
                scope:
                  type: string
                  description: "Optional. Used for 2FA. Format: 'totp:123456'"
              required:
                - username
                - password
      responses:
        '200':
          description: Successful authentication
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Token'
        '401':
          description: Unauthorized
  /2fa/setup:
    get:
      summary: Setup 2FA
      description: Generates a secret and returns a QR code for setting up TOTP-based 2FA.
      tags:
        - 2FA
      security:
        - bearerAuth: []
      responses:
        '200':
          description: A PNG image of the QR code to be scanned by an authenticator app.
          content:
            image/png:
              schema:
                type: string
                format: binary
        '400':
          description: 2FA is already enabled for this user.
  /2fa/verify:
    post:
      summary: Verify and Enable 2FA
      description: Verifies the provided TOTP code and enables 2FA for the user.
      tags:
        - 2FA
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                totp_code:
                  type: string
                  description: The 6-digit code from the authenticator app.
              required:
                - totp_code
      responses:
        '200':
          description: 2FA enabled successfully.
        '400':
          description: Invalid code or 2FA setup not initiated.
  /2fa/disable:
    delete:
      summary: Disable 2FA
      description: Disables TOTP-based 2FA for the currently authenticated user.
      tags:
        - 2FA
      security:
        - bearerAuth: []
      responses:
        '204':
          description: 2FA disabled successfully.
  /me:
    get:
      summary: Get current user info
      tags:
        - Users
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized
  /users/{user_email}/deactivate:
    post:
      summary: Deactivate a user (Admin only)
      tags:
        - Users
      security:
        - bearerAuth: []
      parameters:
        - name: user_email
          in: path
          required: true
          schema:
            type: string
            format: email
      responses:
        '200':
          description: User deactivated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '403':
          description: Forbidden - Admin access required.
        '404':
          description: User not found.

components:
  schemas:
    UserCreate:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
      required:
        - email
        - password
    User:
      type: object
      properties:
        email:
          type: string
          format: email
        is_active:
          type: boolean
          default: true
        is_admin:
          type: boolean
          default: false
        is_two_factor_enabled:
          type: boolean
          default: false
      required:
        - email
    Token:
      type: object
      properties:
        access_token:
          type: string
        token_type:
          type: string
      required:
        - access_token
        - token_type
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT